graph/path/internal: move grid graphs into testgraphs
